Eliminate use of variable-length arrays in prescan.c.
Instead, we precalculate the minimum and maximum length of a label
that *could* be in the table, bail out of validate_charset early if
the label's length is out of range, and then we can use a fixed-length
stack buffer.

I am open to clever ideas for setting UNCANON_LABEL_MIN/MAX with
preprocessor magic instead of an awk script.